introduce webpack to module development
