+++ modules/secure-upload/build.sh
@@ -16,9 +16,9 @@ fi
 # one file that's being deployed.
 v2skylink=$(skynet-utils generate-v2skylink module $seed)
 echo "uploading module to skynet..."
-v1skylink=$(skynet-utils upload-file build/index.js) || (echo "upload failed" && exit 1)
+v1skylink=$(skynet-utils upload-file dist/index.js) || exit 1
 echo "updating module resolver link"
-skynet-utils upload-to-v2skylink $v1skylink module $seed || (echo "v2skylink update failed" && exit 1)
+skynet-utils upload-to-v2skylink $v1skylink module $seed || exit 1
 
 echo "Module can be accessed at: $v2skylink"
 echo $v2skylink > build/module.txt
