+++ modules/secure-upload/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "secure-upload",
+  "version": "0.0.1",
+  "description": "skynet kernel module to securely upload files to skynet",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack --config webpack.config.js --mode=production"
+  },
+  "author": "David Vorick",
+  "license": "MIT",
+  "devDependencies": {
+    "ts-loader": "^9.2.8",
+    "typescript": "^4.6.3",
+    "webpack": "^5.71.0",
+    "webpack-cli": "^4.9.2"
+  }
+}
