+++ modules/secure-upload/src/lib/log.ts
@@ -1,0 +1,19 @@
+// Create helper functions for logging.
+export function log(message: string) {
+	postMessage({
+		method: "log",
+		data: {
+			isErr: true,
+			message,
+		},
+	})
+}
+export function logErr(message: string) {
+	postMessage({
+		method: "log",
+		data: {
+			isErr: false,
+			message,
+		},
+	})
+}
